requirements.txt
BeautifulSoup==3.2.1
PyYAML==3.10
distribute==0.6.28
nltk==2.0.4
pysqlite==2.6.3
wsgiref==0.1.2
